lost policy, updated now, fix up spacing

+++ app/controllers/actions_controller.rb
@@ -6,19 +6,18 @@ class ActionsController < ApplicationController
   end
 
   def index
-    # @actions = policy_scope(Action.all)
-    @actions = Action.order(:name).page params[:page]
-    # authorize @actions
+    @actions = policy_scope(Action.order(:title).page(params[:page]))
+    authorize @actions
   end
 
   def edit
     @action = Action.find(params[:id])
-    # authorize @action
+    authorize @action
   end
 
   def update
     @action = Action.find(params[:id])
-    # authorize @action
+    authorize @action
 
     if @action.update_attributes(permitted_attributes(@action))
       redirect_to action_path, notice: "Action updated"
@@ -29,12 +28,12 @@ class ActionsController < ApplicationController
 
   def show
     @action = Action.find(params[:id])
-    # authorize @action
+    authorize @action
   end
 
   def destroy
     @action = Action.find(params[:id])
-    # authorize @action
+    authorize @action
 
     if @action.destroy
       redirect_to actions_path, notice: "Action deleted"
